# Backend/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "bankApI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Client/AddClientNotification": {
      "post": {
        "tags": [
          "Client"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DCNotifications"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/ManageClients/AddNewClient": {
      "post": {
        "tags": [
          "ManageClients"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DPersonClientS"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/ManageClients/GetClientById{id}": {
      "get": {
        "tags": [
          "ManageClients"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DPersonClientG"
            }
          }
        }
      }
    },
...
    "/api/ManageClients/AddAccount": {
      "post": {
        "tags": [
          "ManageClients"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DBankEmail"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/DAccountGet"
            }
          }
        }
      }
    },
    "/api/ManageClients/AddClientNotification": {
      "post": {
        "tags": [
          "ManageClients"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DCNotifications"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ManageClients/FreezeUnfreezeClientAccount": {
      "put": {
        "tags": [
          "ManageClients"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DSetEmailState"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ManageClients/Deposit": {
      "post": {
        "tags": [
          "ManageClients"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DDepositWithdraw"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/ManageClients/Withdraw": {
      "post": {
        "tags": [
          "ManageClients"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DDepositWithdraw"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "definitions": {
    "DAccount": {
      "type": "object",
      "properties": {
        "passWord": {
          "type": "string"
        },
        "balance": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "DAccountGet": {
      "type": "object",
      "properties": {
        "accountId": {
          "type": "string"
        },
        "balance": {
          "format": "double",
          "type": "number"
        },
        "isFrozen": {
          "type": "boolean"
        },
        "createdAt": {
          "format": "date",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DBankEmail": {
      "type": "object",
      "properties": {
        "currentAccount": {
          "type": "string"
        },
        "passWord": {
          "type": "string"
        },
        "initialBalance": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
    "DCNotifications": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "body": {
          "type": "string"
        },
        "type": {
          "format": "int32",
          "type": "integer"
        },
        "accountId": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "DCardGet": {
      "type": "object",
      "properties": {
        "cardNumber": {
          "type": "string"
        },
        "expirationDate": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DDepositWithdraw": {
      "type": "object",
      "properties": {
        "clientAccount": {
          "type": "string"
        },
        "employeeAccount": {
          "type": "string"
        },
        "amount": {
          "format": "double",
          "type": "number"
        }
      },
      "additionalProperties": false
    },
...
    "DPerson": {
      "required": [
        "address",
        "birthDate",
        "email",
        "firstName",
        "lastName",
        "phoneNumber"
      ],
      "type": "object",
      "properties": {
        "firstName": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "lastName": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "birthDate": {
          "format": "date",
          "type": "string"
        },
        "email": {
          "minLength": 1,
          "type": "string"
        },
        "address": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "phoneNumber": {
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "DPersonClientG": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "birthDate": {
          "format": "date",
          "type": "string"
        },
        "accountInfo": {
          "$ref": "#/definitions/DAccountGet"
        },
        "cardInfo": {
          "$ref": "#/definitions/DCardGet"
        },
        "personalEmail": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "roleType": {
          "type": "string"
        },
        "bankEmails": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DAccountGet"
          }
        },
        "cards": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DCardGet"
          }
        },
        "isActive": {
          "type": "boolean"
        }
      },
      "additionalProperties": false
    },
    "DPersonClientS": {
      "type": "object",
      "properties": {
        "person": {
          "$ref": "#/definitions/DPerson"
        },
        "account": {
          "$ref": "#/definitions/DAccount"
        }
      },
      "additionalProperties": false
    },
    "DSetEmailState": {
      "type": "object",
      "properties": {
        "accountId": {
          "type": "string"
        },
        "state": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
  }
}